Make `get_chunk_voxel_position` public
